Organize the codebase
To improve maintainability + readability across the codebase, organize the code.
